Rename local metric package
