+++ cmd/manager/main.go
@@ -12,8 +12,8 @@ import (
 	"github.com/openshift/aws-account-operator/pkg/apis"
 	"github.com/openshift/aws-account-operator/pkg/controller"
 	"github.com/openshift/aws-account-operator/pkg/credentialwatcher"
-	metrics "github.com/openshift/aws-account-operator/pkg/metrics"
-	custommetrics "github.com/openshift/operator-custom-metrics/pkg/metrics"
+	"github.com/openshift/aws-account-operator/pkg/localmetrics"
+	"github.com/openshift/operator-custom-metrics/pkg/metrics"
 	"github.com/operator-framework/operator-sdk/pkg/leader"
 	"github.com/operator-framework/operator-sdk/pkg/log/zap"
 	sdkVersion "github.com/operator-framework/operator-sdk/version"
@@ -109,18 +109,19 @@ func main() {
 	}
 
 	//Create metrics endpoint and register metrics
-	metricsServer := custommetrics.NewBuilder().WithPort("9090").WithPath("/metrics").
-		WithCollectors(metrics.MetricTotalAWSAccounts).
-		WithCollectors(metrics.MetricTotalAccountCRs).
-		WithCollectors(metrics.MetricTotalAccountCRsUnclaimed).
-		WithCollectors(metrics.MetricTotalAccountCRsClaimed).
-		WithCollectors(metrics.MetricTotalAccountCRsFailed).
-		WithCollectors(metrics.MetricTotalAccountClaimCRs).
-		WithCollectors(metrics.MetricPoolSizeVsUnclaimed).
+	metricsServer := metrics.NewBuilder().WithPort("9090").WithPath("/metrics").
+		WithCollectors(localmetrics.MetricTotalAWSAccounts).
+		WithCollectors(localmetrics.MetricTotalAccountCRs).
+		WithCollectors(localmetrics.MetricTotalAccountCRsUnclaimed).
+		WithCollectors(localmetrics.MetricTotalAccountCRsClaimed).
+		WithCollectors(localmetrics.MetricTotalAccountCRsFailed).
+		WithCollectors(localmetrics.MetricTotalAccountClaimCRs).
+		WithCollectors(localmetrics.MetricPoolSizeVsUnclaimed).
+		WithCollectors(localmetrics.MetricTotalAccountPendingVerification).
 		GetConfig()
 
 	// Configure metrics if it errors log the error but continue
-	if err := custommetrics.ConfigureMetrics(context.TODO(), *metricsServer); err != nil {
+	if err := metrics.ConfigureMetrics(context.TODO(), *metricsServer); err != nil {
 		log.Error(err, "Failed to configure Metrics")
 	}
